+++ local_projects/73bd2a0e-3c4e-4c4b-aea5-2ae2000cea05/generated_code.py
@@ -1,1 +1,0 @@
-print('v2 updated')